# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Simply LaRae Beauty Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface Service {
  id: number;
  slug: string;
  name: string;
  tagline: string;
  description: string;
  price: number;
  priceDisplay: string;
  features: string[];
  isPopular: boolean;
  sortOrder: number;
}

export interface IntakeData {
  fullName: string;
  email: string;
  ageRange: string;
  skinType: string;
  skinTone: string;
  undertone: string;
  sensitivities?: string;
  hairColor?: string;
  eyeColor?: string;
  hasFreckles?: string;
  makeupExperience: string;
  preferredBrands?: string;
  dislikedBrands?: string;
  budgetPreference: string;
  makeupStyle: string;
  desiredCoverage?: string;
  preferredFinish?: string;
  favoriteProducts?: string;
  productsNeverWork?: string;
  everydayGoals?: string;
  eventNeeds?: string;
  biggestFrustration?: string;
  storePreference?: string;
  routinePreference?: string;
  primerPreference?: string;
  applicationMethod?: string;
}

export interface CreateSubmissionRequest {
  serviceId: number;
  intakeData: IntakeData;
  selfieUrl?: string;
  sessionId: string;
}

export type SubmissionStatus =
  (typeof SubmissionStatus)[keyof typeof SubmissionStatus];

export const SubmissionStatus = {
  pending: "pending",
  processing: "processing",
  review: "review",
  completed: "completed",
} as const;

export interface Submission {
  id: number;
  serviceId: number;
  sessionId: string;
  intakeData: IntakeData;
  selfieUrl?: string | null;
  status: SubmissionStatus;
  createdAt: string;
  updatedAt: string;
}

export type ProductPriceRange =
  (typeof ProductPriceRange)[keyof typeof ProductPriceRange];

export const ProductPriceRange = {
  drugstore: "drugstore",
  midrange: "midrange",
  highend: "highend",
} as const;

export interface Product {
  id: number;
  name: string;
  brand: string;
  category: string;
  description: string;
  price: number;
  priceRange: ProductPriceRange;
  imageUrl?: string | null;
  purchaseUrl?: string | null;
  tags: string[];
}

export interface Recommendation {
  id: number;
  submissionId: number;
  product: Product;
  category: string;
  reason: string;
  applicationTip?: string | null;
  sortOrder: number;
}

export interface CartItem {
  id: number;
  sessionId: string;
  product: Product;
  quantity: number;
  addedAt: string;
}

export interface Cart {
  sessionId: string;
  items: CartItem[];
  totalItems: number;
  estimatedTotal: number;
}

export interface AddToCartRequest {
  sessionId: string;
  productId: number;
  quantity: number;
}

export interface ContactRequest {
  name: string;
  email: string;
  subject: string;
  message: string;
}

export interface Faq {
  id: number;
  question: string;
  answer: string;
  category: string;
  sortOrder: number;
}

export type ListProductsParams = {
  category?: string;
  priceRange?: ListProductsPriceRange;
};

export type ListProductsPriceRange =
  (typeof ListProductsPriceRange)[keyof typeof ListProductsPriceRange];

export const ListProductsPriceRange = {
  drugstore: "drugstore",
  midrange: "midrange",
  highend: "highend",
} as const;

export type GetCartParams = {
  sessionId: string;
};