# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Simply LaRae Beauty Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get cart items
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cart items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartRequest"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
  /cart/{id}:
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
        category:
          type: string
        description:
          type: string
        price:
          type: number
        priceRange:
          type: string
          enum: [drugstore, midrange, highend]
        imageUrl:
          type: string
          nullable: true
        purchaseUrl:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - brand
        - category
        - description
        - price
        - priceRange
        - tags
...
    Cart:
      type: object
      properties:
        sessionId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        totalItems:
          type: integer
        estimatedTotal:
          type: number
      required:
        - sessionId
        - items
        - totalItems
        - estimatedTotal
    CartItem:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: string
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        addedAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - product
        - quantity
        - addedAt
    AddToCartRequest:
      type: object
      properties:
        sessionId:
          type: string
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - sessionId
        - productId
        - quantity
...